% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg__fitness_from_model.R
\name{dg__fitness_from_model}
\alias{dg__fitness_from_model}
\title{Predict fitness from computed dG model}
\usage{
dg__fitness_from_model(
  dg_model0,
  variant_data0,
  varlist,
  parlist,
  calc_performance = FALSE,
  per_testset = FALSE,
  overwrite = TRUE,
  bfit_ddg0 = FALSE
)
}
\arguments{
\item{dg_model0}{dG model(s) for which fitness is to be predicted, either long or wide format}

\item{variant_data0}{variant_data data.table}

\item{varlist}{varlist}

\item{parlist}{parlist}

\item{calc_performance}{logical, if Pearson's R should be calculated for predicted versus measured fitness, default = FALSE}

\item{per_testset}{logical, if TRUE, predicts fitness only for test set variants (dg_model0 needs to contain test_set variable and is expected to be in wide format), default = FALSE}

\item{overwrite}{logical, overwrite predicted fitness values if already present in variant_data0?, default = TRUE}

\item{bfit_ddg0}{logical, also calculate binding fitness values assuming folding or binding ddGs are all 0, default = FALSE}
}
\value{
returns a list with a variant_data data.table with predited fitness values, and a prediction_performance data.table if calc_performance == TRUE
}
\description{
Predict fitness from computed dG model
}
