% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_basic_analyses.R
\name{dg_basic_analyses}
\alias{dg_basic_analyses}
\title{Produces some basic analyses plots}
\usage{
dg_basic_analyses(dataset_folder, model_name)
}
\arguments{
\item{dataset_folder}{absolute path to the dataset folder, is created if non-existent}

\item{model_name}{name of the model that should be computed on the dataset

 @import data.table}
}
\description{
Produces some basic analyses plots
}
