% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_bootstrap.R
\name{dg_bootstrap}
\alias{dg_bootstrap}
\title{Bootstrap procedure to estimate the uncertainty of free energies parameters of a model}
\usage{
dg_bootstrap(
  dataset_folder,
  model_name,
  iteration = 1,
  return_model = FALSE,
  maxit = 10000
)
}
\arguments{
\item{dataset_folder}{absolute path to the dataset folder, is created if non-existent}

\item{model_name}{name of the model that should be computed on the dataset}

\item{iteration}{integer, to calucate a set of models and get reproducible results}

\item{return_model}{logical, if TRUE, returns a data.table with fitted values for all model parameters, default = FALSE}

\item{maxit}{integer, maximum number of iterations by optim algorithm, default = 1e4}
}
\value{
writes the boostrapped model parameters as .txt file to $dataset_folder/$model_name/tmp/dg_boostrap_$iteration
}
\description{
Bootstrap procedure to estimate the uncertainty of free energies parameters of a model
}
