% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_collect_models.R
\name{dg_collect_models}
\alias{dg_collect_models}
\title{Reads and collects individual dg models from tmp folder}
\usage{
dg_collect_models(
  dataset_folder,
  model_name,
  model_averaging = "mean",
  which_test_set = 0,
  stage = "model"
)
}
\arguments{
\item{dataset_folder}{absolute path to the dataset folder, is created if non-existent}

\item{model_name}{name of the model that should be computed on the dataset}

\item{model_averaging}{c("mean" [default], "median"), how to average estimated parameters over different train/test sets}

\item{which_test_set}{for 0, the default, reads individual models from training/test sets, for a number greater 0, only reads models from this training/test set}

\item{stage}{c("model" [default], "bootstrap", "pll"), which stage to collect models from, for bootstrap and pll stage, the functions adds their results to the dg_model.txt table}
}
\value{
returns a data.table with fitted values for all model parameters
}
\description{
Reads and collects individual dg models from tmp folder
}
