% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_estimate.R
\name{dg_estimate}
\alias{dg_estimate}
\title{Estimate free energies (dG) for a model from a dataset}
\usage{
dg_estimate(
  dataset_folder,
  model_name,
  which_test_set = 0,
  iteration = 1,
  return_model = FALSE,
  maxit = 10000
)
}
\arguments{
\item{dataset_folder}{absolute path to the dataset folder, is created if non-existent}

\item{model_name}{name of the model that should be computed on the dataset}

\item{which_test_set}{integer, test_set to exclude from model training, default = 0, i.e. will be chosen depending on iteration parameter}

\item{iteration}{integer, to calucate a set of models and get reproducible results}

\item{return_model}{logical, if TRUE, returns a data.table with fitted values for all model parameters, default = FALSE}

\item{maxit}{integer, maximum number of iterations by optim algorithm, default = 1e4}
}
\value{
writes the model parameters as .txt file to $base_folder/$dataset_name/$model_name/tmp/model$iteration
}
\description{
Estimate free energies (dG) for a model from a dataset
}
