% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_model_fn.R
\name{dg_model_fn}
\alias{dg_model_fn}
\title{the dg model function to be minimized}
\usage{
dg_model_fn(par, parlist, varlist)
}
\arguments{
\item{par}{vector of starting parameters for modeling}

\item{parlist}{list with all model parameters}

\item{varlist}{list with all variables}
}
\description{
the dg model function to be minimized
}
