% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_model_gr.R
\name{dg_model_gr}
\alias{dg_model_gr}
\title{gradient for the dg model function to be minimized}
\usage{
dg_model_gr(par, parlist, varlist)
}
\arguments{
\item{par}{vector of starting parameters for modeling}

\item{parlist}{list with all model parameters}

\item{varlist}{list with all variables}
}
\description{
gradient for the dg model function to be minimized
}
