% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_run_parallel.R
\name{dg_run_parallel}
\alias{dg_run_parallel}
\title{A wrapper for parallel exectuion of dg_estimate() or dg_bootstrap()}
\usage{
dg_run_parallel(
  dataset_folder,
  model_name,
  which_test_set = 0,
  iterations,
  Ncores = 8,
  stage = "model"
)
}
\arguments{
\item{dataset_folder}{absolute path to the dataset folder, is created if non-existent}

\item{model_name}{name of the model that should be computed on the dataset}

\item{which_test_set}{integer, (for stage = "model" only): test_set to exclude from model training, default = 0, i.e. will be chosen depending on iteration parameter}

\item{iterations}{a vector of integers to iterate over in dg_estimate(iteration = iterations[i])}

\item{Ncores}{integer, how many cores to use for parallelization}

\item{stage}{"model" or "bootstrap"}
}
\value{
writes the model parameters as .txt file to $dataset_folder/$model_name/tmp/dg_model_$testset_$iteration
}
\description{
A wrapper for parallel exectuion of dg_estimate() or dg_bootstrap()
}
