% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_datasets.R
\name{prepare_datasets}
\alias{prepare_datasets}
\title{Prepare DiMSum-processed doubledeepPCA datasets for free energy estimation}
\usage{
prepare_datasets(
  dataset_folder,
  abundancepca_files,
  bindingpca_files,
  wt_seq,
  train_test_split = c(10, 0.1),
  fitness_scale = "lin"
)
}
\arguments{
\item{dataset_folder}{absolute path to the dataset folder, is created if non-existent}

\item{abundancepca_files}{absolute path to the DiMSum-processed abundancePCA .RData files containing aa_seq, fitness and sigma cols; multiple files will be treated as independent measurements}

\item{bindingpca_files}{absolute path to the DiMSum-processed bindingPCA .RData files containing aa_seq, fitness and sigma cols; multiple files will be used as independent measurements}

\item{wt_seq}{a character string of the wild-type/reference aa sequence}

\item{train_test_split}{c(A, B), Ax test split of variants, where each split has size B; default c(10, 0.1), i.e. 10x cross validation}

\item{fitness_scale}{either "lin" or "log"; for "lin", assumes fitness = [0,Inf], with wild-type fitness 1; for "log", assumes fitness = [-Inf, Inf], with wild-type fitness = 0, default = "lin"}
}
\value{
writes a .RData file to $base_folder/#dataset_name/data/fitness_dataset.RData containing the varlist list with all necessary variables to compute dg models for the dataset
}
\description{
Prepare DiMSum-processed doubledeepPCA datasets for free energy estimation
}
