% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dg_model.R
\name{prepare_dg_model}
\alias{prepare_dg_model}
\title{Prepare parameters for a specific dG model}
\usage{
prepare_dg_model(
  dataset_folder,
  model_name,
  no_folded_states = 1,
  dg_extremes = 15,
  fixed_par = c()
)
}
\arguments{
\item{dataset_folder}{absolute path to the dataset folder, is created if non-existent}

\item{model_name}{name of the model that should be computed on the dataset}

\item{no_folded_states}{integer, '1': one folded, binding-competent state, '2': binding-incompetent and binding-competent folded states, default = 1}

\item{dg_extremes}{float, lower and upper bounds for dG parameters, default = 15}

\item{fixed_par}{list of parameters that are being fixed at their starting values for model fitting}
}
\value{
writes a .RData file to $dataset_folder/$model_name/parameter_list.RData containing the parlist list with all necessary parameters to compute a specific dG model
}
\description{
Prepare parameters for a specific dG model
}
