% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_basic_analyses.R
\name{dg_basic_analyses}
\alias{dg_basic_analyses}
\title{Produces some basic analyses plots}
\usage{
dg_basic_analyses(
  dataset_folder,
  model_name,
  color_type = "",
  max_variants = 3000,
  datasets_ab = c(1, 1)
)
}
\arguments{
\item{dataset_folder}{absolute path to the dataset folder, is created if non-existent}

\item{model_name}{name of the model that should be computed on the dataset}

\item{color_type}{variable with which to color variants; can be any column_name from structural_properties file; any other than the default "" will exclude variants with multiple mutations and the wild-type variant}

\item{max_variants}{maximual number of variants to plot, default = 3e3, will be downsampled if # variants > max_variants}

\item{datasets_ab}{index of datasets to use (in case multiple abundance or binding datasets were supplied)}
}
\description{
Produces some basic analyses plots
}
